+++ packages/orq-rc/src/orq_ai_sdk/models/get_v2_human_eval_setsop.py
@@ -100,7 +100,7 @@ class GetV2HumanEvalSetsResponseBody2(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-17T14:01:25.519Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-18T04:50:35.672Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -185,7 +185,7 @@ class GetV2HumanEvalSetsResponseBody1(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-17T14:01:25.519Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-18T04:50:35.672Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")