+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVAY6VC3GE84CCREE1W2T6J8"
+        "01KVCH33GKH5V48FHGWG8V1XDG"
     )
     r"""The unique identifier of the data source"""