+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVAY6VC5V6P7QZDNRNTJWVQC"
+        "01KVCH33GNJSXP99AMVSTMTDYB"
     )
     r"""The unique identifier of the data source"""