+++ packages/orq-rc/RELEASES.md
@@ -8819,3 +8819,13 @@ Based on:
 - [python v4.11.0-rc.31] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.31] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.31 - packages/orq-rc
+
+## 2026-06-18 04:53:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.782.1 (2.908.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.32] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.32] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.32 - packages/orq-rc